% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bcdf}
\alias{bcdf}
\title{Construct and evaluate a bivariate empirical cumulative distribution function}
\usage{
bcdf(data, eval)
}
\arguments{
\item{data}{A two column matrix for constructing the BECDF}

\item{eval}{A two column matrix for input into the BECDF}
}
\value{
A numeric vector of output values from the BECDF
}
\description{
Construct a bivariate empirical cumulative distribution function (BECDF)
using \code{data} and pass each of the \code{eval} points through the BECDF.
}
\examples{
data(iris)
sample1 <- as.matrix(iris[iris$Species == "virginica", 1:2])
sample2 <- as.matrix(iris[iris$Species == "versicolor", 1:2])

bcdf(sample1, sample2)
}
