% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BivarDistDiffTest.R
\name{BivarDistDiffTest}
\alias{BivarDistDiffTest}
\title{The rotational distDiffR test}
\usage{
BivarDistDiffTest(
  data1,
  data2,
  numRot = 8,
  numPerms = 999,
  psiFun = CalcPsiRWS,
  seedNum = NULL
)
}
\arguments{
\item{data1}{a two column matrix of bivariate observations from one sample}

\item{data2}{a two column matrix of bivariate observations from another sample}

\item{numRot}{an integer number of rotational shifts of the pooled samples}

\item{numPerms}{an integer number of permutations of the original data}

\item{psiFun}{a function specifying the Psi statistic calculation}

\item{seedNum}{an integer random seed value}
}
\value{
A list including three objects:
(1) the Psi statistic computed on the original data
(2) a vector of Psi statistics computed on the permuted data
(3) the p-value for the test
}
\description{
The rotational distDiffR test
}
