% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcGroupPsiCWS.R
\name{CalcGroupPsiCWS}
\alias{CalcGroupPsiCWS}
\title{The Psi_CWS statistic for aggregated group data}
\usage{
CalcGroupPsiCWS(data, groups, subjects)
}
\arguments{
\item{data}{A two column matrix of the bivariate pooled samples}

\item{groups}{A numeric vector of sample (or group) labels (use either 1 or 2)}

\item{subjects}{A numeric vector of subject labels}
}
\value{
the Psi_CWS statistic
}
\description{
This statistic computes the complementary weighted squared (CWS) differences
between the averaged subject empirical cumulative distribution functions for
the two samples.
}
