% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcGroupPsiRWS.R
\name{CalcGroupPsiRWS}
\alias{CalcGroupPsiRWS}
\title{The Psi_RWS statistic for aggregated group data}
\usage{
CalcGroupPsiRWS(data, groups, subjects)
}
\arguments{
\item{data}{A two column matrix of the bivariate pooled samples}

\item{groups}{A numeric vector of sample (or group) labels (use either 1 or 2)}

\item{subjects}{A numeric vector of subject labels}
}
\value{
the Psi_RWS statistic
}
\description{
This statistic computes the reversed weighted squared (RWS) differences between
the averaged subject empirical cumulative distribution functions for the two samples.
}
