% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcPsiCWS.R
\name{CalcPsiCWS}
\alias{CalcPsiCWS}
\title{The Psi_CWS Statistic}
\usage{
CalcPsiCWS(data, subjects)
}
\arguments{
\item{data}{a two column matrix of the bivariate pooled samples}

\item{subjects}{a numerical vector of sample labels (use either 1 or 2)}
}
\value{
the Psi_CWS statistic
}
\description{
This statistic computes the complementary weighted squared (CWS) differences
between the empirical cumulative distribution functions for the two samples.
}
