% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcPsiDWA.R
\name{CalcPsiDWA}
\alias{CalcPsiDWA}
\title{The Psi_DWA Statistic}
\usage{
CalcPsiDWA(data, subjects)
}
\arguments{
\item{data}{a two column matrix of the bivariate pooled samples}

\item{subjects}{a numerical vector of sample labels (use either 1 or 2)}
}
\value{
the Psi_DWA statistic
}
\description{
This statistic computes the double weighted absolute (DWA) differences between
the empirical cumulative distribution functions for the two samples.
}
