% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcPsiRWA.R
\name{CalcPsiRWA}
\alias{CalcPsiRWA}
\title{The Psi_RWA Statistic}
\usage{
CalcPsiRWA(data, subjects)
}
\arguments{
\item{data}{a two column matrix of the bivariate pooled samples}

\item{subjects}{a numerical vector of sample labels (use either 1 or 2)}
}
\value{
the Psi_RWA statistic
}
\description{
This statistic computes the reversed weighted absolute (RWA) differences between
the empirical cumulative distribution functions for the two samples.
}
