% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcPsiRWS.R
\name{CalcPsiRWS}
\alias{CalcPsiRWS}
\title{The Psi_RWS Statistic}
\usage{
CalcPsiRWS(data, subjects)
}
\arguments{
\item{data}{a two column matrix of the bivariate pooled samples}

\item{subjects}{a numerical vector of sample labels (use either 1 or 2)}
}
\value{
the Psi_RWS statistic
}
\description{
This statistic computes the reversed weighted squared (RWS) differences between
the empirical cumulative distribution functions for the two samples.
}
