% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcPsiUWS.R
\name{CalcPsiUWS}
\alias{CalcPsiUWS}
\title{The Psi_UWS Statistic}
\usage{
CalcPsiUWS(data, subjects)
}
\arguments{
\item{data}{a two column matrix of the bivariate pooled samples}

\item{subjects}{a numerical vector of sample labels (use either 1 or 2)}
}
\value{
the Psi_UWS statistic
}
\description{
This statistic computes the uniformly weighted squared (UWS) differences between
the empirical cumulative distribution functions for the two samples.
}
