% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashMat.R
\name{hashMat}
\alias{hashMat}
\title{Assigns a hash value to a two-column matrix.}
\usage{
hashMat(mat)
}
\arguments{
\item{mat}{A two column matrix of bivariate observations.}
}
\value{
A numeric hash value
}
\description{
The \code{hashMat()} function assigns hash numbers to the two-column sample
matrices for the purpose of providing identical test results regardless of
the order in which the input data is passed to the \code{distdiffr()} or
\code{grouped_distdiffr()} functions.
}
