% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{NumToroShiftData}
\alias{NumToroShiftData}
\title{Apply a toroidal shift to the pooled samples using a number of points}
\usage{
NumToroShiftData(data, n1, n2, numShifts)
}
\arguments{
\item{data}{A two column matrix of the pooled samples}

\item{n1}{An integer sample size for the first sample}

\item{n2}{An integer sample size for the second sample}

\item{numShifts}{A numeric number of points to be used as toroidal shift origins}
}
\value{
A two column matrix of the shifted pooled samples
}
\description{
Apply a toroidal shift to the pooled samples using a number of points
}
