% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PropToroShiftData}
\alias{PropToroShiftData}
\title{Apply a toroidal shift to the pooled samples using a proportion of points}
\usage{
PropToroShiftData(data, n1, n2, propPnts = 1)
}
\arguments{
\item{data}{A two column matrix of the pooled samples}

\item{n1}{An integer sample size for the first sample}

\item{n2}{An integer sample size for the second sample}

\item{propPnts}{A numeric proportion of points to be used as toroidal shift origins}
}
\value{
A list of toroidal shifted pooled sample matrices
}
\description{
The \code{PropToroShiftData()} function produces a list of toroidal shifted
versions of the two-column input matrix. The number of toroidal shifts
is (the ceiling of) the proportion (\code{propPnts}) multiplied by the combined
sample size. The origins of the toroidal shifts are randomly selected from
the combined samples. The pooled \code{data} is assumed to list all of the first
sample of size \code{n1} before the second sample (of size \code{n2}).
}
\examples{
data(iris)
sample1 <- as.matrix(iris[iris$Species == "setosa", 1:2])
sample2 <- as.matrix(iris[iris$Species == "virginica", 1:2])
pooled_data <- rbind(sample1, sample2)
n1 <- nrow(sample1)
n2 <- nrow(sample2)

# Creates a list of 0.1 times (n1 + n2) = 10 toroidal shifts of the pooled data
output <- PropToroShiftData(pooled_data, n1, n2, 0.1)
summary(output)
}
