% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RotateData.R
\name{RotateData}
\alias{RotateData}
\title{Create rotated versions of the data}
\usage{
RotateData(data, numRotations)
}
\arguments{
\item{data}{A two column matrix of the bivariate combined samples}

\item{numRotations}{A non-negative integer specifying the number of rotations to be applied to the data within 360 degrees.}
}
\value{
A list of matrices containing the coordinates for each version of the rotated data (including the original data, which is the first matrix of the list)
}
\description{
This function produces a list of rotated versions of the two-column input
matrix. Specifically, the number of rotations (i.e., an integer passed to
\code{numRotations}) divides a complete circle into \code{numRotations} equal angles,
and \code{numRotations} rotated versions of the input data are output in a list.
}
\examples{
data(iris)
sample1 <- as.matrix(iris[iris$Species == "setosa", 1:2])

# Generate five rotated versions of sample1 (every 72 degrees) within 360 degrees.
RotateData(sample1, 5)
}
