% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ToroShiftData}
\alias{ToroShiftData}
\title{Apply a toroidal shift to the pooled samples}
\usage{
ToroShiftData(data, n1, n2, propPnts = 1)
}
\arguments{
\item{data}{A two column matrix of the pooled samples}

\item{n1}{An integer sample size for the first sample}

\item{n2}{An integer sample size for the second sample}

\item{propPnts}{A numeric proportion of points to be used as toroidal shift origins}
}
\value{
A two column matrix of the shifted pooled samples
}
\description{
Apply a toroidal shift to the pooled samples
}
