% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregatedRotateToroDiffTest.R
\name{AggregatedRotateToroDiffTest}
\alias{AggregatedRotateToroDiffTest}
\title{The combined rotational and toroidal shift distDiffR test for aggregated group data}
\usage{
AggregatedRotateToroDiffTest(
  aggdata1,
  aggdata2,
  numRot = 8,
  propPnts = NULL,
  numShifts = NULL,
  numPerms = 99,
  psiFun = CalcGroupPsiRWS,
  seedNum = NULL
)
}
\arguments{
\item{aggdata1}{A three column matrix of bivariate observations from one sample with the third column being the subject labels}

\item{aggdata2}{A three column matrix of bivariate observations from another sample with the third column being the subject labels}

\item{numRot}{An integer number of rotational shifts of the pooled samples}

\item{propPnts}{A numeric proportion of points to be used as toroidal shift origins}

\item{numShifts}{A numeric integer. The number of points to be used as toroidal shift origins. Must be less than the pooled sample size.}

\item{numPerms}{An integer number of permutations of the original data}

\item{psiFun}{A function specifying the Psi statistic calculation. Default is the CalcGroupPsiRWS.}

\item{seedNum}{An integer random seed value}
}
\value{
A list including three objects:
(1) the Psi statistic computed on the original data
(2) a vector of Psi statistics computed on the permuted data
(3) the p-value for the test
}
\description{
The combined rotational and toroidal shift distDiffR test for aggregated group data
}
