% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bcdf}
\alias{bcdf}
\title{Construct and compute bivariate cumulative distribution function values}
\usage{
bcdf(data, eval)
}
\arguments{
\item{data}{A two column matrix for constructing the empirical bivariate cumulative distribution function (EBCDF)}

\item{eval}{A two column matrix for input into the EBCDF}
}
\value{
A numeric vector of output values from the EBCDF
}
\description{
Construct and compute bivariate cumulative distribution function values
}
