% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcPsiCWA.R
\name{CalcPsiCWA}
\alias{CalcPsiCWA}
\title{The Psi_CWA Statistic}
\usage{
CalcPsiCWA(data, subjects)
}
\arguments{
\item{data}{a two column matrix of the bivariate pooled samples}

\item{subjects}{a numerical vector of sample labels (use either 1 or 2)}
}
\value{
the Psi_CWA statistic
}
\description{
This statistic computes the complementary weighted absolute (CWA) differences
between the empirical cumulative distribution functions for the two samples.
}
