% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcPsiDWA.R
\name{CalcPsiDWA}
\alias{CalcPsiDWA}
\title{The Psi DWA Statistic}
\usage{
CalcPsiDWA(data, subjects)
}
\arguments{
\item{data}{A two column matrix of the bivariate pooled samples}

\item{subjects}{A numerical vector of sample labels (use either 1 or 2)}
}
\value{
The Psi DWA statistic
}
\description{
This statistic computes the double weighted absolute (DWA) differences between
the empirical cumulative distribution functions for the two samples.
For more information, see \insertCite{mckinney2022extensions;textual}{distdiffR}
and \insertCite{mckinney2021extensions;textual}{distdiffR}.
}
\examples{
data(iris)
pooled_data <- iris[iris$Species \%in\% c("setosa", "virginica"), 1:2]
sample_labels <- rep(1:2, c(sum(iris$Species == "setosa"),
                            sum(iris$Species == "virginica")))

CalcPsiDWA(as.matrix(pooled_data), sample_labels)
}
\references{
\insertRef{mckinney2022extensions}{distdiffR}

\insertRef{mckinney2021extensions}{distdiffR}
}
