% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcPsiDWS.R
\name{CalcPsiDWS}
\alias{CalcPsiDWS}
\title{The Psi_DWS Statistic}
\usage{
CalcPsiDWS(data, subjects)
}
\arguments{
\item{data}{a two column matrix of the bivariate pooled samples}

\item{subjects}{a numerical vector of sample labels (use either 1 or 2)}
}
\value{
the Psi_DWS statistic
}
\description{
This statistic computes the double weighted squared (DWS) differences between
the empirical cumulative distribution functions for the two samples.
}
