% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcPsiUWA.R
\name{CalcPsiUWA}
\alias{CalcPsiUWA}
\title{The Psi_UWA Statistic}
\usage{
CalcPsiUWA(data, subjects)
}
\arguments{
\item{data}{a two column matrix of the bivariate pooled samples}

\item{subjects}{a numerical vector of sample labels (use either 1 or 2)}
}
\value{
the Psi_UWA statistic
}
\description{
This statistic computes the uniformly weighted absolute (UWA) differences between
the empirical cumulative distribution functions for the two samples.
}
