% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_distdiffr.R
\name{grouped_distdiffr}
\alias{grouped_distdiffr}
\title{The combined rotational and toroidal shift distdiffR test for aggregated group data}
\usage{
grouped_distdiffr(
  aggdata1,
  aggdata2,
  numRot = 8,
  propPnts = NULL,
  numShifts = NULL,
  shiftThrshld = 25,
  numPerms = 999,
  psiFun = CalcGroupPsiCWS,
  seedNum = NULL
)
}
\arguments{
\item{aggdata1}{A three column matrix of bivariate observations from one sample with the third column being the numeric subject labels}

\item{aggdata2}{A three column matrix of bivariate observations from another sample with the third column being the numeric subject labels}

\item{numRot}{An integer number of rotational shifts of the pooled samples}

\item{propPnts}{A numeric proportion of points to be used as toroidal shift origins. Cannot provide both propPnts and numShifts. If neither are provided, shiftThrshld is used.}

\item{numShifts}{A numeric integer. The number of points to be used as toroidal shift origins. Must be less than the pooled sample size. Cannot provide both propPnts and numShifts. If neither are provided, shiftThrshld is used.}

\item{shiftThrshld}{A numeric integer. Used if neither propPnts or numShifts are provided. If the pooled sample size is less than shiftThrshld, every point will be used as a toroidal shift origin. Otherwise, only a random sample of shiftThrshld points will be used.}

\item{numPerms}{An integer number of permutations of the original data}

\item{psiFun}{A function specifying the Psi statistic calculation. Default is the CalcGroupPsiCWS.}

\item{seedNum}{An integer random seed value}
}
\value{
A list including three objects:
(1) the Psi statistic computed on the original data
(2) a vector of Psi statistics computed on the permuted data
(3) the p-value for the test
}
\description{
The \code{grouped_distdiffr()} function conducts two-sample permutation tests of
distributional equality based on differences in the bivariate empirical
cumulative density functions (BECDFs). The differences in BECDFs are computed
across a series of rotations, toroidal shifts, or both rotations and toroidal
shifts of the combined data (specified via \code{testType}). The number of rotations
and toroidal shifts may be specified (via \code{numRot} or \code{numShifts},
respectively). The number of toroidal shifts may also be determined by a
proportion of the combined sample size (via \code{propPnts}).
}
\details{
Additionally, \code{grouped_distdiffr()} assumes multiple sources (subjects) are
contributing to each sample. As such, the function weights each sources
contribution as to treat each equally within the samples, respectively.
However, this test is only currently available with the grouped CWS statistic
and employs both rotational and toroidal shifts. For more information, see
\insertCite{mckinney2022extensions;textual}{distdiffR} and
\insertCite{mckinney2021extensions;textual}{distdiffR}.
}
\examples{
# Randomly assign all three species to two samples
# The species serve as subject labels within each sample
seedNum <- 123
set.seed(seedNum)

data(iris)
irisPermuted <- iris
irisPermuted$Species <- rep(1:3, each = 50)
irisPermuted <- irisPermuted[sample.int(nrow(irisPermuted)), ]
sample1 <- as.matrix(irisPermuted[1:75, c(1:2, 5)])
sample2 <- as.matrix(irisPermuted[76:150, c(1:2, 5)])

output <- grouped_distdiffr(sample1,
                            sample2,
                            seedNum = seedNum)
output$pval
}
\references{
\insertRef{mckinney2022extensions}{distdiffR}

\insertRef{mckinney2021extensions}{distdiffR}
}
