% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashMat.R
\name{hashMat}
\alias{hashMat}
\title{Assigns a hash value to a two-column matrix.}
\usage{
hashMat(mat)
}
\arguments{
\item{mat}{A two column matrix of bivariate observations.}
}
\value{
A numeric hash value
}
\description{
Assigns a hash value to a two-column matrix.
}
