% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{NumToroShiftData}
\alias{NumToroShiftData}
\title{Apply a toroidal shift to the pooled samples using a number of points}
\usage{
NumToroShiftData(data, n1, n2, numShifts)
}
\arguments{
\item{data}{A two column matrix of the pooled samples}

\item{n1}{An integer sample size for the first sample}

\item{n2}{An integer sample size for the second sample}

\item{numShifts}{A numeric number of points to be used as toroidal shift origins}
}
\value{
A list of toroidal shifted pooled sample matrices
}
\description{
The \code{NumToroShiftData()} function produces a list of toroidal shifted
versions of the two-column input matrix. The number of toroidal shifts
is an integer passed to \code{numShifts}. The origins of the toroidal shifts are
randomly selected from the combined samples. The pooled \code{data} is assumed to
list all of the first sample of size \code{n1} before the second sample (of size
\code{n2}).
}
\examples{
data(iris)
sample1 <- as.matrix(iris[iris$Species == "setosa", 1:2])
sample2 <- as.matrix(iris[iris$Species == "virginica", 1:2])
pooled_data <- rbind(sample1, sample2)
n1 <- nrow(sample1)
n2 <- nrow(sample2)

# Create a list of five toroidal shifts of the pooled data
output <- NumToroShiftData(pooled_data, n1, n2, 25)
summary(output)
}
