% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RotateData.R
\name{RotateData}
\alias{RotateData}
\title{Create rotated versions of the data}
\usage{
RotateData(data, numRotations)
}
\arguments{
\item{data}{a two column matrix of the bivariate pooled samples}

\item{numRotations}{a non-negative integer specifying the number of rotations to be applied to the data within 360 degrees.}
}
\value{
a list of matrices containing the coordinates for each version of the rotated data (including the original data)
}
\description{
This function creates rotated versions of the pooled samples
}
